+++ src/akkudoktoreos/prediction/pvforecast.py
@@ -7,7 +7,12 @@ from pydantic import Field, computed_field, field_validator, model_validator
 from akkudoktoreos.config.configabc import SettingsBaseModel
 from akkudoktoreos.core.coreabc import get_prediction
 from akkudoktoreos.prediction.pvforecastabc import PVForecastProvider
+from akkudoktoreos.prediction.pvforecastforecastsolar import (
+    PVForecastForecastSolarCommonSettings,
+)
 from akkudoktoreos.prediction.pvforecastimport import PVForecastImportCommonSettings
+from akkudoktoreos.prediction.pvforecastpvnode import PVForecastPVNodeCommonSettings
+from akkudoktoreos.prediction.pvforecastsolcast import PVForecastSolcastCommonSettings
 from akkudoktoreos.prediction.pvforecastvrm import PVForecastVrmCommonSettings
 
 
@@ -18,7 +23,14 @@ def pvforecast_provider_ids() -> list[str]:
     except:
         # Prediction may not be initialized
         # Return at least provider used in example
-        return ["PVForecastAkkudoktor", "PVForecastImport", "PVForecastVrm"]
+        return [
+            "PVForecastAkkudoktor",
+            "PVForecastImport",
+            "PVForecastVrm",
+            "PVForecastPVNode",
+            "PVForecastForecastSolar",
+            "PVForecastSolcast",
+        ]
 
     return [
         provider.provider_id()
@@ -178,6 +190,18 @@ class PVForecastCommonProviderSettings(SettingsBaseModel):
     PVForecastVrm: Optional[PVForecastVrmCommonSettings] = Field(
         default=None,
         json_schema_extra={"description": "PVForecastVrm settings", "examples": [None]},
+    )
+    PVForecastPVNode: Optional[PVForecastPVNodeCommonSettings] = Field(
+        default=None,
+        json_schema_extra={"description": "PVForecastPVNode settings", "examples": [None]},
+    )
+    PVForecastForecastSolar: Optional[PVForecastForecastSolarCommonSettings] = Field(
+        default=None,
+        json_schema_extra={"description": "PVForecastForecastSolar settings", "examples": [None]},
+    )
+    PVForecastSolcast: Optional[PVForecastSolcastCommonSettings] = Field(
+        default=None,
+        json_schema_extra={"description": "PVForecastSolcast settings", "examples": [None]},
     )